+++ src/components/homepage-parts/index.js
@@ -1,0 +1,8 @@
+export { Intro } from './intro'
+export { Work } from './work'
+export { Writing } from './writing'
+export { Speaking } from './speaking'
+export { OpenSource } from './open-source'
+export { Featured } from './featured'
+export { Testimonials } from './testimonials'
+export { Contact } from './contact'
