+++ src/components/homepage-parts/work.js
@@ -1,17 +1,18 @@
 import React from 'react'
 
-export default () => (
+const Work = () => (
   <>
     <h2>Trusted By</h2>
     <p>
-      I have been working with big and small companies doing WordPress and React. Here are some of them.
+      I have been working with big and small companies doing WordPress and
+      React. Here are some of them.
     </p>
     <section className="projects-container">
       <div className="work">
         <a href="https://www.rtcamp.com/">
           <img
             className="work-logo"
-            src={require('../images/rtcamp-logo-black.svg')}
+            src={require('../../images/rtcamp-logo-black.svg')}
             alt="rtCamp"
           />
         </a>
@@ -23,7 +24,7 @@ export default () => (
         <a href="https://www.capbase.com/">
           <img
             className="work-logo"
-            src={require('../images/capbase.png')}
+            src={require('../../images/capbase.png')}
             alt="Capbase"
           />
         </a>
@@ -35,7 +36,7 @@ export default () => (
         <a href="https://www.learndash.com/">
           <img
             className="work-logo"
-            src={require('../images/LearnDash.png')}
+            src={require('../../images/LearnDash.png')}
             alt="LearnDash"
           />
         </a>
@@ -47,7 +48,7 @@ export default () => (
         <a href="https://javascriptforwp.com/">
           <img
             className="work-logo"
-            src={require('../images/jsforwp.png')}
+            src={require('../../images/jsforwp.png')}
             alt="JS for WP"
           />
         </a>
@@ -59,31 +60,33 @@ export default () => (
         <a href="https://www.simplur.com/">
           <img
             className="work-logo"
-            src={require('../images/simplur.svg')}
+            src={require('../../images/simplur.svg')}
             alt="Simplur"
           />
         </a>
         <div className="project-technologies">
-          <span>React</span> <span>Gatsby</span> <span>WordPress</span> <span>WooCommerce</span>{' '}
+          <span>React</span> <span>Gatsby</span> <span>WordPress</span>{' '}
+          <span>WooCommerce</span>{' '}
         </div>
       </div>
       <div className="work">
         <a href="http://octaware.com/">
           <img
             className="work-logo"
-            src={require('../images/octaware_tech_logo_1.jpg')}
+            src={require('../../images/octaware_tech_logo_1.jpg')}
             alt="Octaware Technologies"
           />
         </a>
         <div className="project-technologies">
-          <span>WordPress</span> <span>WooCommerce</span> <span>Integrations</span> {' '}
+          <span>WordPress</span> <span>WooCommerce</span>{' '}
+          <span>Integrations</span>{' '}
         </div>
       </div>
       <div className="work">
         <a href="https://www.jatstudio.com/">
           <img
             className="work-logo"
-            src={require('../images/jatstudio.svg')}
+            src={require('../../images/jatstudio.svg')}
             alt="Jatstudio"
           />
         </a>
@@ -94,3 +97,5 @@ export default () => (
     </section>
   </>
 )
+
+export { Work }
